spritesheet iterator
